Move behave fixtures to datastacks
