+++ datastacks/datastacks/constants.py
@@ -5,9 +5,13 @@ AZURE_RESOURCE_GROUP_NAME = os.environ.get("AZURE_RESOURCE_GROUP_NAME")
 AZURE_DATA_FACTORY_NAME = os.environ.get("AZURE_DATA_FACTORY_NAME")
 AZURE_REGION_NAME = os.environ.get("AZURE_REGION_NAME")
 AZURE_STORAGE_ACCOUNT_NAME = os.environ.get("AZURE_STORAGE_ACCOUNT_NAME")
+AZURE_CONFIG_ACCOUNT_NAME = os.environ.get("AZURE_CONFIG_ACCOUNT_NAME")
 
 RAW_CONTAINER_NAME = "raw"
 ADLS_URL = f"https://{AZURE_STORAGE_ACCOUNT_NAME}.dfs.core.windows.net"
+CONFIG_CONTAINER_NAME = "config"
+CONFIG_BLOB_URL = f"https://{AZURE_CONFIG_ACCOUNT_NAME}.blob.core.windows.net"
+
 AUTOMATED_TEST_OUTPUT_DIRECTORY_PREFIX = "automated_test"
 
 CONFIG_CONTAINER_NAME = "config"
