+++ de_templates/ingest/Ingest_SourceType_SourceName/tests/end_to_end/features/environment.py.jinja
@@ -3,7 +3,7 @@ from behave import use_fixture
 from behave.model import Scenario
 from behave.runner import Context
 
-from fixtures import azure_adls_clean_up
+from datastacks.behave.fixtures import azure_adls_clean_up
 
 SQL_DB_INGEST_DIRECTORY_NAME = "{{ pipeline_name }}"
